Now we exit the app when the main window closes
This is to keep the broadcast via NDI obviously tied to the app - plus we had no way to reopen the window if we closed it, so exiting the app seemed reasonable. In the future we may revisit so that the window can be reopened. But then we'd have to restrict to one window, etc. So that would be more complex.